Move isWorkspaceEditLike from core/capability-probes to refactor/workspace-edit
isWorkspaceEditLike is a domain-specific type guard for WorkspaceEdit, a class
that lives in src/refactor/src/workspace-edit.ts. Its previous home in
src/core/src/utils/capability-probes.ts (alongside truly generic probes like
isRegExpLike and isMapLike) violated the core/refactor boundary: core should
hold only cross-domain utilities, not guards for refactor-domain types.

Changes:
- Remove isWorkspaceEditLike from src/core/src/utils/capability-probes.ts
- Remove the explicit isWorkspaceEditLike override from CoreNamespace in
  src/core/src/index.ts (no longer needed; Core no longer exposes this)
- Add isWorkspaceEditLike to src/refactor/src/workspace-edit.ts, co-located
  with the WorkspaceEdit class it guards
- Update src/refactor/src/refactor-engine.ts to import isWorkspaceEditLike
  directly from ./workspace-edit.js instead of Core.isWorkspaceEditLike
- Move the test from src/core/test/capability-probes.test.ts to
  src/refactor/test/workspace-edit.test.ts, where it belongs
- Fix pre-existing TypeScript errors in gml-ast-builder.ts by widening
  getTokenStartIndex and buildDirectiveKeywordRange to accept number inputs
  (ctx?.start returns number | Token via TypeScript widening)
