+++ src/parser/src/ast/gml-ast-builder.ts
@@ -244,9 +244,13 @@ export default class GameMakerASTBuilder {
         return tokenRecord;
     }
 
-    private getTokenStartIndex(token: Token | ParserToken | null | undefined): number | null {
-        if (!token) {
+    private getTokenStartIndex(token: number | Token | ParserToken | null | undefined): number | null {
+        if (!token && token !== 0) {
             return null;
+        }
+
+        if (typeof token === "number") {
+            return Number.isFinite(token) ? token : null;
         }
 
         if (typeof token.start === "number") {
@@ -261,7 +265,7 @@ export default class GameMakerASTBuilder {
     }
 
     private buildDirectiveKeywordRange(
-        token: Token | ParserToken | null | undefined,
+        token: number | Token | ParserToken | null | undefined,
         keyword: DirectiveKeyword
     ): DirectiveKeywordRange | null {
         const start = this.getTokenStartIndex(token);
