+++ src/refactor/test/workspace-edit.test.ts
@@ -5,7 +5,7 @@
 import assert from "node:assert/strict";
 import { test } from "node:test";
 
-import { getWorkspaceArrays, WorkspaceEdit } from "../src/workspace-edit.js";
+import { getWorkspaceArrays, isWorkspaceEditLike, WorkspaceEdit } from "../src/workspace-edit.js";
 
 void test("getWorkspaceArrays extracts valid arrays from workspace", () => {
     const workspace = new WorkspaceEdit();
@@ -94,4 +94,33 @@ void test("getWorkspaceArrays preserves array contents", () => {
 
     assert.deepEqual(metadataEdits, expectedMetadata);
     assert.deepEqual(fileRenames, expectedRenames);
+});
+
+void test("isWorkspaceEditLike identifies valid workspace-edit-like objects", () => {
+    const validWorkspaceEdit = {
+        edits: [],
+        addEdit() {},
+        groupByFile() {
+            return new Map();
+        }
+    };
+
+    assert.equal(isWorkspaceEditLike(validWorkspaceEdit), true);
+    assert.equal(isWorkspaceEditLike(new WorkspaceEdit()), true);
+});
+
+void test("isWorkspaceEditLike rejects non-conforming objects", () => {
+    assert.equal(isWorkspaceEditLike({ edits: [] }), false);
+    assert.equal(isWorkspaceEditLike({ edits: [], addEdit() {} }), false);
+    assert.equal(isWorkspaceEditLike({ addEdit() {}, groupByFile() {} }), false);
+    assert.equal(isWorkspaceEditLike(null), false);
+    assert.equal(isWorkspaceEditLike(), false);
+    assert.equal(
+        isWorkspaceEditLike({
+            edits: "not an array",
+            addEdit() {},
+            groupByFile() {}
+        }),
+        false
+    );
 });
